PHPUnit_Successfull_Install_and_Test
PHPUnit was successfully installed on local main directory, so the PHPUnit can be removed from the prototype. This was performed in attempt to  mimic how the repo was being tested on Travis. This way develoment of unit tests can be performed locally without having to push and have them done on Travis.